uv: Upgrade to ad382bca

+++ deps/uv/src/ares/ares_send.c
@@ -77,7 +77,7 @@ void ares_send(ares_channel channel, const unsigned char *qbuf, int qlen,
     }
 
   /* Compute the query ID.  Start with no timeout. */
-  query->qid = (unsigned short)DNS_HEADER_QID(qbuf);
+  query->qid = DNS_HEADER_QID(qbuf);
   query->timeout.tv_sec = 0;
   query->timeout.tv_usec = 0;
